recipe create

+++ Services/Interfaces/IService.cs
@@ -3,7 +3,7 @@ namespace WebCooking.Services.Interfaces;
 public interface IService<T> where T : class
 {
     Task<IEnumerable<T>> GetAllAsync();
-    Task<T> GetByIdAsync(int id);
+    Task<T> GetByIdAsync(long id);
     Task AddAsync(T entity);
     Task UpdateAsync(T entity);
     Task DeleteAsync(int id);
